+++ src/module/vue/components/character-sheet-tabs/sf-legacies.vue
@@ -16,6 +16,7 @@
         v-for="(progressItem, i) in starredProgresses"
         :key="`progress-item-${i}`"
         :item="progressItem"
+        :show-star="true"
       />
     </section>
   </article>